backend/requirements.txt
# Web framework
fastapi==0.115.5
slowapi==0.1.9             # Rate limiting
uvicorn[standard]==0.32.1
gunicorn==23.0.0
python-multipart==0.0.12  # File uploads

# Database
psycopg2-binary==2.9.11
alembic==1.14.1

# Auth & encryption
bcrypt==5.0.0
cryptography==46.0.3      # Fernet for Todoist token encryption
itsdangerous==2.2.0       # HMAC-signed tokens for email links

# Caching
cachetools==5.5.0

# Config
pydantic-settings==2.7.0
pydantic[email]==2.10.4
python-dotenv==1.0.1

# HTTP & scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# LLM (OpenAI-compatible client for local LLM)
openai==1.57.4

# Todoist
todoist-api-python==2.1.7

# Image processing
Pillow==11.0.0

# Task queue (replaces APScheduler for distributed, multi-worker scheduling)
celery[redis]==5.4.0
redis==5.2.1

# Gmail OAuth (same deps as calendar_bot)
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.156.0

# Web search for recipe ideas
tavily-python==0.5.0       # Primary: Tavily (requires API key, free tier limited)
duckduckgo-search==6.3.7   # Fallback: DDG (no API key, no credit limits)
